move icon to assets folder
